# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ShiftManager API
  version: "1.0.0"
  description: REST API for the ShiftManager club volunteer-hours system.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MemberWrite:
      type: object
      properties:
        firstName:  { type: string }
        lastName:   { type: string }
        email:      { type: string, format: email }
        isActive:   { type: boolean }
        role:       { $ref: '#/components/schemas/MemberRole' }
        joinedAt:   { type: string, format: date-time }
        individualGoalHours: { type: number, nullable: true }
...
    MeResponse:
      type: object
      required: [id, role, email]
      properties:
        id:    { $ref: '#/components/schemas/UUID' }
        role:  { $ref: '#/components/schemas/MemberRole' }
        email: { type: string, format: email }
    TokenResponse:
      type: object
      required: [token, expiresIn]
      properties:
        token:     { type: string, description: Freshly signed JWT }
        expiresIn: { type: integer, description: Seconds until the new token expires }

    # ── Event ─────────────────────────────────────────────────────────────────
    EventStatus:
      type: string
      enum: [draft, published, cancelled, completed]
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []
paths:
...
  /auth/callback:
    get:
      operationId: authCallback
      tags: [auth]
      summary: OIDC authorization-code callback
      security: []
      parameters:
        - { name: code, in: query, required: false, schema: { type: string } }
        - { name: state, in: query, required: false, schema: { type: string } }
      responses:
        "302": { description: Redirect to SPA with token in fragment }
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Return the currently authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Re-issue a JWT with a fresh expiry for the current session (A-004)
      responses:
        "200":
          description: Freshly issued token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Invalidate the session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ── Members ─────────────────────────────────────────────────────────────────
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: isActive, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Member' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createMember
      tags: [members]
      summary: Create a member (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberWrite' }
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /members/import:
    post:
      operationId: importMembers
      tags: [members]
      summary: Import members from CSV (Vorstand+)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /members/export:
    get:
      operationId: exportMembers
      tags: [members]
      summary: Export members as CSV
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string, format: binary }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /members/me/preferences:
    put:
      operationId: updateMyPreferences
      tags: [members]
      summary: Update own reminder opt-out preference
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberPreferences' }
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberPreferences' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /members/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getMember
      tags: [members]
      summary: Get a single member
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateMember
      tags: [members]
      summary: Update a member (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberWrite' }
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deactivateMember
      tags: [members]
      summary: Deactivate (soft-delete) a member (Vorstand+)
      responses:
        "204": { description: Deactivated }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /members/{id}/export-data:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: exportMemberData
      tags: [members]
      summary: GDPR data export (own data or Vorstand+)
      responses:
        "200":
          description: Export JSON
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /members/{id}/gdpr-delete:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: gdprDeleteMember
      tags: [members]
      summary: Anonymize a member (GDPR, Vorstand+)
      responses:
        "200":
          description: Anonymized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── Events ──────────────────────────────────────────────────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/EventStatus' } }
        - { name: visibility, in: query, schema: { $ref: '#/components/schemas/EventVisibility' } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Event' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventWrite' }
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /events/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event with full shift timeline
      responses:
        "200":
          description: Event timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventTimeline' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventWrite' }
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (Veranstaltungsleiter+, draft/cancelled only)
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /events/{id}/publish:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: publishEvent
      tags: [events]
      summary: Publish a draft event (Veranstaltungsleiter+)
      responses:
        "200":
          description: Published event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /events/{id}/timeline:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getEventTimeline
      tags: [events]
      summary: Get just the shift timeline for an event
      responses:
        "200":
          description: Event timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventTimeline' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
  /events/{id}/attachments:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: listEventAttachments
      tags: [events]
      summary: List files attached to an event (V-008)
      responses:
        "200":
          description: Attachments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EventAttachment' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: uploadEventAttachment
      tags: [events]
      summary: Upload an image or document attachment for an event (Veranstaltungsleiter+, V-008)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Uploaded attachment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventAttachment' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "415": { description: Unsupported file type }
  /events/{id}/attachments/{attachmentId}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
      - { name: attachmentId, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    delete:
      operationId: deleteEventAttachment
      tags: [events]
      summary: Delete an event attachment (Veranstaltungsleiter+, V-008)
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /events/{eventId}/shifts:
    parameters:
      - { name: eventId, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: createShift
      tags: [shifts]
      summary: Add a shift to an event (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShiftWrite' }
      responses:
        "201":
          description: Created shift
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shift' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── Shifts ───────────────────────────────────────────────────────────────────
  /shifts/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    put:
      operationId: updateShift
      tags: [shifts]
      summary: Update a shift (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShiftWrite' }
      responses:
        "200":
          description: Updated shift
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shift' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift (Veranstaltungsleiter+)
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /shifts/{id}/register:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: registerShift
      tags: [shifts]
      summary: Register the authenticated member for a shift
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterShiftRequest' }
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "409": { description: Already registered }
    delete:
      operationId: deregisterShift
      tags: [shifts]
      summary: Cancel the authenticated member's registration
      parameters:
        - { name: registrationId, in: query, required: true, schema: { $ref: '#/components/schemas/UUID' } }
        - { name: force, in: query, schema: { type: boolean } }
      responses:
        "204": { description: Deregistered }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /shifts/{id}/confirm:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    post:
      operationId: confirmShiftRegistration
      tags: [shifts]
      summary: Confirm a registration (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [registrationId]
              properties:
                registrationId: { $ref: '#/components/schemas/UUID' }
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── Hours ───────────────────────────────────────────────────────────────────
  /hours/me:
    get:
      operationId: getMyHours
      tags: [hours]
      summary: Get the authenticated member's hour account (active club year)
      responses:
        "200":
          description: Hour account with entries
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberHourAccount' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /hours/account:
    get:
      operationId: getMemberAccount
      tags: [hours]
      summary: Get a member's hour account by memberId+clubYearId query params
      parameters:
        - { name: memberId, in: query, required: true, schema: { $ref: '#/components/schemas/UUID' } }
        - { name: clubYearId, in: query, required: true, schema: { $ref: '#/components/schemas/UUID' } }
      responses:
        "200":
          description: Hour account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberHourAccount' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /hours/confirm:
    post:
      operationId: confirmShiftHours
      tags: [hours]
      summary: Record confirmed hours after a shift (Veranstaltungsleiter+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmShiftHoursRequest' }
      responses:
        "201":
          description: Hour entry created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HourEntry' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /hours/summary:
    get:
      operationId: getYearSummary
      tags: [hours]
      summary: Year-end hour summary for all members (Veranstaltungsleiter+)
      parameters:
        - { name: clubYearId, in: query, required: true, schema: { $ref: '#/components/schemas/UUID' } }
      responses:
        "200":
          description: Summary rows
          content:
            application/json:
              schema: { type: array, items: { type: object } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /hours/manual:
    post:
      operationId: manualBooking
      tags: [hours]
      summary: Create a manual hour entry (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualBookingRequest' }
      responses:
        "201":
          description: Hour entry created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HourEntry' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /hours/{memberId}:
    parameters:
      - { name: memberId, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    get:
      operationId: getMemberHours
      tags: [hours]
      summary: Get any member's hour account (active club year)
      responses:
        "200":
          description: Hour account with entries
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberHourAccount' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /hours/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { $ref: '#/components/schemas/UUID' } }
    put:
      operationId: correctHourEntry
      tags: [hours]
      summary: Correct an hour entry (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hours, description]
              properties:
                hours:       { type: number }
                description: { type: string }
      responses:
        "200":
          description: Corrected entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HourEntry' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteHourEntry
      tags: [hours]
      summary: Delete an hour entry (Vorstand+)
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── Stats ────────────────────────────────────────────────────────────────────
  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: System-wide statistics (Veranstaltungsleiter+)
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SystemStats' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

  # ── Settings ─────────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get application settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings (Vorstand+)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppSettings' }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
...
  /kiosk/events:
    get:
      operationId: listKioskEvents
      tags: [kiosk]
      summary: List all published events (no auth)
      security: []
      responses:
        "200":
          description: Published events
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Event' } }
        "403": { description: Kiosk locked }
...